Implemented a basic chat server
There is now a server-wide goroutine that is responsible for keeping track of connected clients and sending messages to all of them. A message is read by the client's reader goroutine, sent to the global chat goroutine, which then sends it to all clients' writer goroutines. We're getting closer to a "real" server architecture.